Separate components and policies in folders
