+++ apps/hotel/src/restaurant/components/directives/view-mode.directive.ts
